Fix time_range attributes types

+++ hashcode/hashcode/time_range.hpp
@@ -2,12 +2,11 @@
 
 
 class time_range{
-
 	private:
-		unsigned int m_start_time;
-		unsigned int m_end_time;
+		unsigned long int m_start_time;
+		unsigned long int m_end_time;
 	public:
-		time_range(unsigned int, unsigned int);
+		time_range(unsigned long int, unsigned long int);
 		~time_range();
 		time_range& operator=(const time_range& time_range);
 		time_range(const time_range& time_range);
